Refactor project, add redux
