+++ package.json
@@ -22,10 +22,10 @@
     "jquery": "^3.5.1",
     "material-table": "^1.59.0",
     "query-string": "^6.12.1",
-    "react": "^16.4.1",
+    "react": "^16.13.1",
     "react-bootstrap": "^1.0.0-beta.17",
-    "react-dom": "^16.4.1",
-    "react-redux": "^5.0.7",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.1",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "redux": "^4.0.5",
