+++ src/components/about-us/AboutUsComponent.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import { Carousel } from 'react-bootstrap';
-import NavbarUnimorph from './NavbarComponent';
-import Footer from './Footer';
 
 export default function AboutUsSection() {
     return (
