+++ src/components/core/Footer.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import styled from 'styled-components'
 
 function Footer() {
     return (
