+++ src/constants.js
@@ -6,7 +6,7 @@ const prod = {
 
 const dev = {
   url: {
-    API_URL: 'http://127.0.0.1:8000/api'
+    API_URL: 'http://unimorph.ethz.ch/api'
   }
 }
